Moving MetaData classes to com.ibm.ws.jpa.management
