--- dev/com.ibm.ws.jpa.container.core/src/com/ibm/ws/jpa/app/package-info.java
@@ -1,14 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2024 IBM Corporation and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License 2.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-2.0/
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- * IBM Corporation - initial API and implementation
- *******************************************************************************/
-@org.osgi.annotation.versioning.Version("1.0.0")
-package com.ibm.ws.jpa.app;
