+++ dev/com.ibm.ws.jpa.container/bnd.bnd
@@ -1,5 +1,5 @@
 #*******************************************************************************
-# Copyright (c) 2017, 2024 IBM Corporation and others.
+# Copyright (c) 2017, 2022 IBM Corporation and others.
 # All rights reserved. This program and the accompanying materials
 # are made available under the terms of the Eclipse Public License 2.0
 # which accompanies this distribution, and is available at
@@ -50,7 +50,6 @@ Include-Resource: \
 
 Export-Package: \
   com.ibm.ws.jpa, \
-  com.ibm.ws.jpa.app, \
   com.ibm.ws.jpa.management, \
   com.ibm.ws.jpa.container.osgi.jndi
