+++ dev/com.ibm.ws.jpa.container/src/com/ibm/ws/jpa/container/osgi/internal/JPAComponentImpl.java
@@ -74,11 +74,11 @@ import com.ibm.ws.jpa.JPAExPcBindingContextAccessor;
 import com.ibm.ws.jpa.JPAProviderIntegration;
 import com.ibm.ws.jpa.JPAPuId;
 import com.ibm.ws.jpa.JPAVersion;
-import com.ibm.ws.jpa.app.ApplicationComponentMetaData;
 import com.ibm.ws.jpa.container.osgi.jndi.JPAJndiLookupInfo;
 import com.ibm.ws.jpa.container.osgi.jndi.JPAJndiLookupInfoRefAddr;
 import com.ibm.ws.jpa.container.osgi.jndi.JPAJndiLookupObjectFactory;
 import com.ibm.ws.jpa.management.AbstractJPAComponent;
+import com.ibm.ws.jpa.management.ApplicationComponentMetaData;
 import com.ibm.ws.jpa.management.JPAApplInfo;
 import com.ibm.ws.jpa.management.JPAEMFPropertyProvider;
 import com.ibm.ws.jpa.management.JPAIntrospection;
